Test/planet endpoint

+++ src/routes/Planets.ts
@@ -139,17 +139,37 @@ router.put("/planets/:id", async (req: Request, res: Response) => {
 
     try {
         const id: number = parseInt(req.params.id);
+
+        const planet: Planet = await repository.findOneBy({ id: id });
+        if (!planet) {
+            throw new ResourceNotFound(
+                "Planet com ID fornecido não foi encontrado.",
+            );
+        }
+
+        const { name, weather, terrain, population, stellarSystemId, ...rest } =
+            req.body;
+        if (Object.keys(rest).length > 0) {
+            throw new InvalidRequestBody(
+                "Corpo da requisição inválido. planet deve conter somente 'name', 'weather', 'terrain', 'population' e 'stellarSystemId'.",
+            );
+        }
+
         await repository.update({ id: id }, req.body);
 
         res.status(HTTPStatus.OK).send({
             status: "success",
-            message: "Requisição processada com sucesso",
-            data: await repository.findOneByOrFail({ id: id }),
+            message: "Planet específicado alterado com sucesso.",
+            data: {
+                planet: await repository.findOneByOrFail({ id: id }),
+            },
         });
-    } catch {
-        res.status(HTTPStatus.INTERNAL_SERVER_ERROR).send({
+    } catch (error) {
+        res.status(
+            error.HTTPStatusCode || HTTPStatus.INTERNAL_SERVER_ERROR,
+        ).send({
             status: "error",
-            message: "Erro ao processar a requisição",
+            message: error.message || "Erro ao processar a requisição",
             data: {},
         });
     }
@@ -160,20 +180,24 @@ router.delete("/planets/:id", async (req: Request, res: Response) => {
 
     try {
         const id: number = parseInt(req.params.id);
-        const planet: Planet = await repository.findOneByOrFail({ id: id });
+        const planet: Planet = await repository.findOneBy({ id: id });
+        if (!planet) {
+            throw new ResourceNotFound("Planet com ID fornecido não existe.");
+        }
+
         await repository.delete({ id: id });
 
         res.status(HTTPStatus.OK).send({
             status: "success",
-            message: "Requisição processada com sucesso",
-            data: {
-                planet,
-            },
+            message: "Planet especificado deletado com sucesso.",
+            data: {},
         });
-    } catch {
-        res.status(HTTPStatus.INTERNAL_SERVER_ERROR).send({
+    } catch (error) {
+        res.status(
+            error.HTTPStatusCode || HTTPStatus.INTERNAL_SERVER_ERROR,
+        ).send({
             status: "error",
-            message: "Erro ao processar a requisição",
+            message: error.message || "Erro ao processar a requisição",
             data: {},
         });
     }
